kanie-lab-agent/backend/pyproject.toml
[project]
name = "kanie-lab-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "firebase-admin>=6.0.0",
    "google-cloud-firestore>=2.19.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
    "claude-agent-sdk",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
anyio_backend = "asyncio"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.4.0",
    "httpx>=0.27.0",
]
